Muevo helpers de strings a módulo aparte
