+++ ckanext/gobar_theme/helpers/__init__.py
@@ -8,3 +8,4 @@ from .cron import *
 from .distribution import *
 from .dates import *
 from .config import *
+from .strings import *
